Reset to base domain
