+++ app/chat/agent.py
@@ -1,0 +1,6 @@
+import abc
+
+class AbstractChatAgent(abc.ABC):
+    @abc.abstractmethod
+    async def execute_flow(self, question: str, conversation: conv_models.ConversationHistory) -> state_models.ChatState:
+        pass
