+++ app/entrypoints/fastapi.py
@@ -6,7 +6,6 @@ import uvicorn
 
 from app import config
 from app.common import mongo, tracing
-from app.conversation_history import router as conversation_history_router
 from app.health import router as health_router
 from app.v2_chat import router as chat_v2_router
 
@@ -35,7 +34,6 @@ app.add_middleware(tracing.TraceIdMiddleware)
 # Setup Routes
 app.include_router(health_router.router)
 app.include_router(chat_v2_router.router, prefix="/v2")
-app.include_router(conversation_history_router.router)
 
 
 def main() -> None:
