--- app/v2_chat/prompts/final_answer.txt
@@ -1,23 +1,0 @@
-Reformat the given text into structured Markdown suitable for GOV.UK webpages. Preserve all factual information without adding new content.
-
-CRITICAL: Start your response immediately with the content. Do not begin with any conversational phrases, introductory statements, or meta-commentary such as:
-- "Here is the content reformatted..."
-- "The following information..."
-- "Based on the text..."
-- "Here are the details..."
-- Any similar introductory phrases
-
-Content requirements:
-1. Always start with a brief overview paragraph that summarises the key information (do not label it as "Overview:" or "Summary:")
-2. Present detailed information as concise bullet points (1-2 sentences each)
-3. Never use headings or subheadings
-4. Use clear, professional British English appropriate for government audiences
-5. Avoid meta-commentary or self-referential phrases
-6. Do not change any information or add new details beyond reformatting
-
-Output format requirements:
-1. Your response must be in Markdown format only
-2. Format bullet points using Markdown syntax (-) when presenting bulleted / list information
-3. Output only the formatted content without explanatory text
-4. Do not use headings (#, ##, ###, etc.)
-5. Begin immediately with the substantive content
