--- app/v2_chat/prompts/hallucination_grader.txt
@@ -1,2 +1,0 @@
-You are a grader assessing whether an llm generation is grounded in /supported by a set of documents.
-Give a binary score of 'yes' or 'no'. 'yes' means that the answer is grounded in / supported by the set of facts.
