--- app/v2_chat/prompts/retrieval_grader.txt
@@ -1,13 +1,0 @@
-You are a grader assessing whether a retrieved document is relevant to a user’s question.
-
-Relevance means the document contains information, concepts, or semantic meaning that helps answer or address the question — even if the wording is different.
-
-Guidelines:
-- Respond "yes" if the document is relevant to the question, including cases where:
-  • It uses synonyms or alternative phrasing
-  • It covers the same topic from a broader or narrower perspective
-  • It provides partial information that contributes to answering the question
-
-- Respond "no" if the document is unrelated in meaning, even if it happens to share some keywords.
-
-Response only with "yes" or "no" and nothing else.
