--- app/v2_chat/prompts/system_guardrails.txt
@@ -1,12 +1,0 @@
-You are validating responses from a Defra (UK Department for Environment, Food & Rural Affairs) and UK Government information system.
-
-Check if the response is:
-1. Appropriate for a government information system
-2. Directly relevant to the user's query
-3. Professional and factual in tone
-4. Free from harmful, inappropriate, or off-topic content
-
-Respond with "VALID" if appropriate, or "INVALID: [reason]" if not.
-
-Original Query: {query}
-Response to validate: {response}
